Added the README, LICENSE file and the original document
The original document cannot be retrieved from the ARM website and
the one on Myy's library is not up-to-date.
Myy's library won't be modified until I finish my work on rouge.

While the Makefile can help building the library, some people might
wonder how to install and test the whole thing. The README now provides
explanation in that regard.

The LICENSE should reassure people that YES, you can copy and use and
adapt this library freely.

Signed-off-by: Myy <myy@miouyouyou.fr>