Making the assembly building part easier for Windows users
Installing a Sourceware AS toolset capable of generating ARMv7 machine
code on Windows is quite a daunting task...
On such platforms, the best option will generally be the the Android
NDK.

While it is entirely possible to locate the cross-compiling toolset
used by the currently installed Android NDK and use it directly in the
Makefile, this might generate more problems that it'll try to solve.

Therefore, an Android.mk file has been added and the ARMv7 Assembly
part of the project can now be built with 'ndk-build', the building
script of the Android NDK.
Another option would be to use CMake, which might be the better
option at long term given the number of useless files 'ndk-build'
require to build the project correctly...

Also, 'ndk-build' uses the only ARMv7 Assembler that does not
understand '@' comments. Therefore, they've been replaced by '//'
comments.

Anyway, Windows users can now build the project more easily. However,
I did not try to build the entire project with Android Studio only...

Signed-off-by: Myy <myy@miouyouyou.fr>